builder-next: Move exporter wrapper to exporter package
