+++ builder/builder-next/worker/containerdworker.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	mobyexporter "github.com/docker/docker/builder/builder-next/exporter"
-	"github.com/docker/docker/builder/builder-next/exporter/overrides"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/exporter"
 	"github.com/moby/buildkit/session"
@@ -33,7 +32,7 @@ func (w *ContainerdWorker) Exporter(name string, sm *session.Manager) (exporter.
 		if err != nil {
 			return nil, err
 		}
-		return overrides.NewExporterWrapper(exp)
+		return mobyexporter.NewWrapper(exp)
 	default:
 		return w.Worker.Exporter(name, sm)
 	}
